use duke-ds-project-transfer instead of delivery
Switching to transfer instead of delivery since it is more up to date
and allows us to view all appropriate transfers (to and from current user).
For the most part leaves the route and components alone.
Passes around a transfer object instead of a delivery.
Only odd part is this results in a route that is /deliveries/<transfer_id>.
This is a trade off which keeps the url consistent with the UI.

Removed un-necessary column from delivery list page.