Adds select filtering for transfer state
Unfortunately the filtering is case sensitive. So to match capitalized choices in the select box with lowercase data values I mirrored the presentation logic from the format-delivery-state helper.
I tried importing that helper and calling it directly, but this seems to require quite a bit of factory injection and ownership.
Seemed overcomplicated to do all that instead of  a `capitalize()` call, so here we are.