computes showProjectDetails for delivery-details
Determining whether or not to show the details is based on the current user and the fields of the transfer. So this has been added to the base can-resend-controller.

The property shall be true for delivery senders (since we assume they'll usually retain view/download access to a project) and true for recipients that have accepted the delivery.